improved unit test coverage
